# server/docs/openapi.yaml
openapi: 3.1.0
info:
  title: grilling-sleek API
  description: REST API for the grilling-sleek platform
  license:
    name: ''
  version: 0.2.0
paths:
...
  /v1/sessions:
    post:
      tags:
      - sessions
      summary: POST /v1/sessions — Create a new session with first round.
      operationId: create_session
      requestBody:
        content:
          application/json:
            schema: {}
        required: true
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSessionResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Max sessions reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CreateSessionResponse:
      type: object
      required:
      - session_id
      - url
      - status
      - current_round
      - created_at
      - expires_at
      properties:
        created_at:
          type: string
        current_round:
          type: integer
          format: int64
        description:
          type:
          - string
          - 'null'
        expires_at:
          type: string
        name:
          type:
          - string
          - 'null'
        session_id:
          type: string
        status:
          type: string
        url:
          type: string
    ErrorResponse:
      type: object
      required:
      - message
      - status
      properties:
        message:
          type: string
        status:
          type: integer
          format: int64
...